+++ docs/CHANGE_LOG.md
@@ -3,4 +3,4 @@
 ## Version 1.2.0:
 - Improved the ignore check to use actual boolean value.
 - Added improved checks around the elements (incase they are removed from the DOM).
-- 
+- Added a new configuration option called "assignMissingIds" (defaults to false), which will assign Ids to the elments that do not contain them.